successfully refactored, NEW WORKING VERSION
